feat: check CHANGELOG for feature
