+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+# Changelog
+
+## [0.1.4] - 2023-05-27
+
+### Added
+
+- Feature A: Added Rwanda verification.
+- Feature B: Added `GET` request like `getBankCodes()`, `getWalletBalance()` e.t.c.
+
+## [0.1.0] - 2023-03-15
+
+### Added
+
+- Initial release of the library.
