+++ example/.gitignore
@@ -22,7 +22,3 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
-
-.cache
-.parcel-cache
-.env
