+++ src/types.ts
@@ -1,0 +1,5 @@
+export type Config = {
+  x_api_key: string;
+  app_id: string;
+  environment: string;
+};
